+++ src/Telemetry/Opt_In_Template.php
@@ -14,7 +14,7 @@ class Opt_In_Template implements Template {
 
 	protected function get_args(): array {
 		return apply_filters( 'stellarwp/telemetry/optin_args', [
-			'plugin_logo'        => plugin_dir_url( __DIR__ ) . 'public/logo.png',
+			'plugin_logo'        => Admin_Resources::get_asset_path() . 'resources/images/stellar-logo.svg',
 			'plugin_logo_width'  => 151,
 			'plugin_logo_height' => 32,
 			'plugin_logo_alt'    => 'StellarWP Logo',
