+++ src/resources/css/styles.css
@@ -1,0 +1,195 @@
+.stellarwp-telemetry {
+	font-family: 'SF Pro Text', 'Helvetica Neue', 'Helvetica', 'Arial', sans-serif;
+	font-size: 14px;
+	font-weight: 400;
+	line-height: 18px;
+	color: #4E4E4E;
+}
+
+.stellarwp-telemetry h1 {
+	font-family: 'SF Pro Text', 'Helvetica Neue', 'Helvetica', 'Arial', sans-serif;
+	font-size: 18px;
+	font-weight: 500;
+	line-height: 24px;
+	letter-spacing: -0.02em;
+	color: #000;
+	margin: 1rem 0;
+}
+
+.stellarwp-telemetry .intro {
+	font-size: 12px;
+	font-weight: 400;
+	line-height: 18px;
+	letter-spacing: -0.01em;
+	text-align: center;
+}
+
+.stellarwp-telemetry ul.questions {
+	text-align: left;
+	width: 85%;
+	margin: 1rem auto;
+}
+
+.stellarwp-telemetry ul.questions li {
+	font-size: 14px;
+	font-weight: 400;
+	line-height: 18px;
+	letter-spacing: -0.01em;
+	text-align: left;
+	margin-bottom: 0.85rem;
+	display: flex;
+	padding: 0 1rem;
+	border-radius: 4px;
+}
+
+.stellarwp-telemetry ul.questions li.active {
+	background: rgb(238 238 238 / 45%);
+	padding: 1rem;
+}
+
+.stellarwp-telemetry ul.questions li:last-child {
+	margin-bottom: 0;
+}
+
+.stellarwp-telemetry ul.questions li input[type="radio"] {
+	margin-top: 1px;
+	margin-right: 1rem;
+}
+
+.stellarwp-telemetry ul.questions li label {
+	width: 100%;
+}
+
+.stellarwp-telemetry ul.questions li textarea {
+	width: 100%;
+	height: 55px;
+	border: 1px solid #DFDFDF;
+	border-radius: 4px;
+	padding: 0.5rem;
+	margin-top: 0.5rem;
+	display: none;
+}
+
+.stellarwp-telemetry ul.questions li.active textarea {
+	display: block;
+}
+
+.stellarwp-telemetry .btn-primary,
+.stellarwp-telemetry .btn-grey,
+.stellarwp-telemetry .btn-text {
+	padding: 8px 12px;
+	background: #0047FF;
+	border-radius: 4px;
+	color: #fff;
+	text-decoration: none;
+	border: 0;
+	outline: none;
+	transition: all 0.1s ease-in-out;
+	cursor: pointer;
+}
+
+.stellarwp-telemetry .btn-primary:hover {
+	background: #0032b7;
+}
+
+.stellarwp-telemetry .btn-grey {
+	background: #eee;
+	color: #000;
+}
+
+.stellarwp-telemetry .btn-grey:hover {
+	background: #ddd;
+}
+
+.stellarwp-telemetry.exit-interview {
+	position: fixed;
+	visibility: hidden;
+	pointer-events: none;
+	opacity: 0;
+	bottom: 0;
+	right: 0;
+	width: 100%;
+	height: 100%;
+	background: rgba(0, 0, 0, 0.5);
+	z-index: 999999;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	transition: all 0.3s ease-in-out;
+}
+
+.stellarwp-telemetry.exit-interview.active {
+	visibility: visible;
+	pointer-events: all;
+	opacity: 1;
+}
+
+.stellarwp-telemetry.exit-interview .stellarwp-telemetry.modal {
+	position: absolute;
+	z-index: 999999;
+	display: flex;
+	flex-direction: column;
+	padding: 32px;
+	width: 90%;
+	top: unset;
+	left: unset;
+	max-width: 585px;
+	background: #fff;
+	box-shadow: 0 0 32px rgba(0, 0, 0, 0.1);
+	border-radius: 4px;
+	text-align: center;
+}
+
+.stellarwp-telemetry.exit-interview .stellarwp-telemetry img.plugin-logo {
+	margin: auto;
+}
+
+.stellarwp-telemetry.wrapper {
+	position: fixed;
+	bottom: 0;
+	right: 0;
+	width: 100%;
+	height: 100%;
+	background: rgba(0, 0, 0, 0.5);
+	z-index: 999999;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	transition: all 0.3s ease-in-out;
+}
+
+.stellarwp-telemetry.modal {
+	position: absolute;
+	z-index: 999999;
+	display: flex;
+	flex-direction: column;
+	padding: 32px;
+	width: 90%;
+	top: 47px;
+	left: 32px;
+	background: #fff;
+	box-shadow: 0 0 32px rgba(0, 0, 0, 0.1);
+	border-radius: 4px;
+}
+
+.stellarwp-telemetry .btn-text {
+	background: #fff;
+	color: #000;
+}
+
+.stellarwp-telemetry .links a:hover,
+.stellarwp-telemetry .btn-text:hover {
+	text-decoration: underline;
+}
+
+.stellarwp-telemetry .links {
+	display: flex;
+	gap: 16px;
+	font-size: 12px;
+	line-height: 16px;
+}
+
+.stellarwp-telemetry .links a {
+	color: #0047FF;
+	text-decoration: none;
+}
