+++ src/views/optin.php
@@ -3,107 +3,8 @@
 /** @var array $args */
 ?>
 <style>
-	.stellarwp-telemetry.wrapper {
-		position: fixed;
-		bottom: 0;
-		right: 0;
-		width: 100%;
-		height: 100%;
-		background: rgba(0, 0, 0, 0.5);
-		z-index: 999999;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		transition: all 0.3s ease-in-out;
-	}
 
-	.stellarwp-telemetry.modal {
-		position: absolute;
-		z-index: 999999;
-		display: flex;
-		flex-direction: column;
-		padding: 32px;
-		width: 90%;
-		top: 47px;
-		left: 32px;
-		background: #fff;
-		box-shadow: 0 0 32px rgba(0, 0, 0, 0.1);
-		border-radius: 4px;
-	}
-
-	.stellarwp-telemetry h1 {
-		font-family: 'SF Pro Text', 'Helvetica Neue', 'Helvetica', 'Arial', sans-serif;
-		font-size: 18px;
-		font-weight: 500;
-		line-height: 24px;
-		letter-spacing: -0.02em;
-		color: #000;
-	}
-
-	.stellarwp-telemetry {
-		font-family: 'SF Pro Text', 'Helvetica Neue', 'Helvetica', 'Arial', sans-serif;
-		font-size: 14px;
-		font-weight: 400;
-		line-height: 18px;
-		color: #4E4E4E;
-	}
-
-	.stellarwp-telemetry .btn-primary,
-	.stellarwp-telemetry .btn-text {
-		padding: 6px 12px;
-		background: #0047FF;
-		border-radius: 4px;
-		color: #fff;
-		text-decoration: none;
-		border: 0;
-		outline: none;
-		transition: all 0.1s ease-in-out;
-		cursor: pointer;
-	}
-
-	.stellarwp-telemetry .btn-primary:hover {
-		background: #0032b7;
-	}
-
-	.stellarwp-telemetry .btn-text {
-		background: #fff;
-		color: #000;
-	}
-
-	.stellarwp-telemetry .links a:hover,
-	.stellarwp-telemetry .btn-text:hover {
-		text-decoration: underline;
-	}
-
-	.stellarwp-telemetry .links {
-		display: flex;
-		gap: 16px;
-		font-size: 12px;
-		line-height: 16px;
-	}
-
-	.stellarwp-telemetry .links a {
-		color: #0047FF;
-		text-decoration: none;
-	}
 </style>
-
-
-<script>
-	// JS document on ready event
-	document.addEventListener("DOMContentLoaded", function () {
-		let wrapper = document.getElementById("modal-wrapper");
-
-		document.getElementById("close-modal").addEventListener("click", function (event) {
-			event.preventDefault();
-			close_modal(wrapper);
-		});
-	});
-
-	function close_modal(wrapper) {
-		wrapper.parentNode.removeChild(wrapper);
-	}
-</script>
 
 <div id="modal-wrapper" class="stellarwp-telemetry wrapper">
 	<section class="stellarwp-telemetry modal">
